refactor: reduce the number of components required to render a timeline item (via props)
