+++ framework/index.js
@@ -9,6 +9,7 @@ import {
   AActivityTimeline,
   AActivityTimelineItem,
   AActivityTimelineItemBody,
+  AActivityTimelineItemContainer,
   AActivityTimelineItemHeader,
   AActivityTimelineItemSubtitle,
   AActivityTimelineItemTags,
@@ -161,6 +162,7 @@ export {
   AActivityTimeline,
   AActivityTimelineItem,
   AActivityTimelineItemBody,
+  AActivityTimelineItemContainer,
   AActivityTimelineItemHeader,
   AActivityTimelineItemSubtitle,
   AActivityTimelineItemTags,
